Set the suite status based on the test results
A Github action can only fail/succeed. Current, if any of the tests
has failed, we mark the whole PR as failed.

Here, we try to be a bit smarter. The conclusion is based on if we
have errors or errors. If there's a an error, red failure mark for
each patch. Dito for waring,

Also, only update the suite if it's in state success/failed.

Signed-off-by: Björn Töpel <bjorn@rivosinc.com>